Restructure frontend to Feature-Sliced Design
