+++ frontend/src/app/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
-import taskReducer from './slices/taskSlice';
-import authReducer from './slices/authSlice';
-import categoryReducer from './slices/categorySlice';
+import taskReducer from '../entities/task/taskSlice';
+import authReducer from '../entities/auth/authSlice';
+import categoryReducer from '../entities/category/categorySlice';
 
 export default configureStore({
   reducer: {
