+++ frontend/src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-import store from './store'
+import './app/index.css';
+import App from './app/App';
+import reportWebVitals from './shared/config/reportWebVitals';
+import store from './app/store'
 import { Provider } from 'react-redux'
 import { BrowserRouter } from "react-router-dom";
