+++ frontend/src/shared/api/authHeader.js
