+++ frontend/src/widgets/FooterBar/FooterBar.js
